Fix cumulative import of bibs for comps (#141)
* Fix cumulative import of bibs for comps

The old code accumulated bibs into a map and injected all bindings from
the map into the database upon import. The new code avoids that by : 1)
adding bibs to the database when theya re parsed (instead of using the
dat in the map) and 2) clearing the map when importing new bibs.

* add test

---------

Co-authored-by: Xavier DUQUESNE <duquesne.xavier.13@gmail.com>